Add elevation chart marker
